*the tones module now caches the waves it generates in the waves directory as "_tone%sHZ%sMS.wav" where the first %s is the pitch in HZ and the second %s is the length in milliseconds. However to avoid filling up disc space, the files are deleted each time the tones module is loaded (i.e when NVDA starts). *insert+s used to turn speech on and off, but now it toggles between 3 speech modes: talk, beeps and off. Talk is just normal speech, off is off, but beeps, is an idea that came from the Dos screen reader, ASAP. This mode plays a 5ms 10000HZ tone (very high and very short) each time an item of speech is supposed to happen. It is very responsive, and it allows you to deal with the situation when you are doing something that is constantly changing the screen, as in millions of files are scrolling up a dos console, or perhaps you are running a setup program that is constantly telling you what file it is copying etc. If you switch to the  beeps speech mode, you only have to wait until the flood of beeps stop and you know the task is finished. This is much faster than having to keep pressing control, or waiting for all the speech to finish. *Shortened the length of progress bar value beeps from 100MS to 40MS. *Changed the speakNewText function for dos consoles to collect together all new peaces of text and send them to the speech synth all in one go (separated by spaces). Previously each time a peace of new text was identified it was sent.

+++ source/tones.py
@@ -5,28 +5,31 @@ import os
 import struct
 import math
 
-sampleRate=11025
+sampleRate=22050.0
 slopeRatio=0.01
-amplitude=10000
+amplitude=5000.0
+
+#Remove old tones
+[os.remove("waves/%s"%x) for x in filter(lambda x: x.startswith("_tone"),os.listdir("waves"))]
 
 def beep(hz,length):
-	fileName=tempfile.mktemp('.wav')
-	waveFile=wave.open(fileName,'w')
-	waveFile.setsampwidth(2)
-	waveFile.setframerate(sampleRate)
-	waveFile.setnchannels(1)
-	volume=0
-	sampleLength=int(length*(sampleRate/1000.0))
-	slopeLength=sampleLength*slopeRatio
-	riseEnd=slopeLength
-	fallStart=sampleLength-slopeLength
-	for sampleCount in xrange(sampleLength):
-		if sampleCount<=riseEnd:
-			volume=sampleCount/slopeLength
-		elif sampleCount>=fallStart:
-			volume=(sampleLength-sampleCount)/slopeLength
-		data=struct.pack('h',amplitude*math.sin((sampleCount*math.pi*2)/(sampleRate/hz))*volume)
-		waveFile.writeframes(data)
-	waveFile.close()
-	winsound.PlaySound(fileName,winsound.SND_FILENAME)
-	os.remove(fileName)
+	fileName="waves\\_tone%sHZ%sMS.wav"%(hz,length)
+	if not os.path.isfile(fileName):
+		waveFile=wave.open(fileName,'w')
+		waveFile.setsampwidth(2)
+		waveFile.setframerate(sampleRate)
+		waveFile.setnchannels(1)
+		volume=0
+		sampleLength=int(length*(sampleRate/1000.0))
+		slopeLength=sampleLength*slopeRatio
+		riseEnd=slopeLength
+		fallStart=sampleLength-slopeLength
+		for sampleCount in xrange(sampleLength):
+			if sampleCount<=riseEnd:
+				volume=sampleCount/slopeLength
+			elif sampleCount>=fallStart:
+				volume=(sampleLength-sampleCount)/slopeLength
+			data=struct.pack('h',amplitude*math.sin((sampleCount*math.pi*2)/(sampleRate/hz))*volume)
+			waveFile.writeframes(data)
+		waveFile.close()
+	winsound.PlaySound(fileName,winsound.SND_FILENAME|winsound.SND_ASYNC)
